+++ test/testutils/getDate.js
@@ -1,3 +1,3 @@
-const getDate = (minOffset) => new Date(new Date().getTime() + (minOffset * 1000 * 60)).toISOString();
+const getDate = (minOffset, baseDate) => new Date((baseDate || new Date()).getTime() + (minOffset * 1000 * 60)).toISOString();
 
 module.exports = getDate;