+++ CHANGELOG.md
@@ -14,6 +14,7 @@ Allowed Types of change: `Added`, `Changed`, `Deprecated`, `Removed`, `Fixed`, `
  - OPS-1297 - Added Changelog github action
  - Unit Tests for event parser
  - fix docker tag latest push
+ - Add integration test for event route and remove skipped tests
 
 ### Changed